+++ wok_code/wok_code/egg-info/history.rst
@@ -3,6 +3,7 @@
 
 * [IMP] ssh.py: -m -s switches accept path with user and host
 * [IMP] deploy_odoo: new property status to display
+* [NEW] do_git_checkout_new_branch.py
 
 2.0.5 (2023-01-13)
 ~~~~~~~~~~~~~~~~~~